Moved octave handling in InputKeyboardNote from GetNoteFromKeypress to separate functions (feel stupid for not figuring this out earlier)

+++ src/audiooscillator.cc
@@ -60,7 +60,7 @@ void AudioOscillator::Release(int index)
 
 void AudioOscillator::Trigger(float freq, float force, int index)
 {
-    noteMap[index] = {force, freq * (float)pow(1.0594630943592953f, semitoneTuning), shallSustain, 0.0f, 0};
+    noteMap[index] = {force, freq * (float)pow(1.0594630943592953f, (float)semitoneTuning), shallSustain, 0.0f, 0};
 }
 
 bool AudioOscillator::Triggered(int index)
